Prometheus Grafana commit
Added the prometheus and grafana installation instructions to the README. Included the updated docker-compose and prometheus yaml files for the installation. Changed the kong.yaml to move the prometheus plugin to global scope. Included all screen shots used for the tutorial under /plugins/images. Updated the docker-compose installation yaml to include STATUS_LISTEN IP and port for the container.